portfolio_backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
whitenoise
gunicorn
dj-database-url
psycopg2-binary
pillow
